Hooks Improvements
Created a test to check if the AndroidManifest.xml was already being patched in config.xml and some other minor changes
