+++ plugin.xml
@@ -48,7 +48,7 @@
 
     <!-- android -->
     <platform name="android">
-        <hook type="after_plugin_install" src="hooks/Android/MABSVersionChecker.js" />
+        <hook type="after_plugin_install" src="hooks/AndroidManifestPatcher.js" />
         <config-file target="res/xml/config.xml" parent="/*">
             <feature name="Camera">
                 <param name="android-package" value="org.apache.cordova.camera.CameraLauncher"/>
